refactor: CalendarOptionsFragment.kt has been renamed to CoursesVisibilityFragment.kt.
refactor: DataChip.kt has removed as it was unused.
